requirements.txt
telebot
rembg
pillow